log_api/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv